workers/pyproject.toml
[project]
name = "sourcebridge-worker"
version = "0.1.0"
description = "SourceBridge AI/ML worker — LLM reasoning, embedding, linking, and requirement ingestion"
requires-python = ">=3.12"
dependencies = [
    "grpcio>=1.62.0",
    "grpcio-tools>=1.62.0",
    "grpcio-health-checking>=1.62.0",
    "grpcio-reflection>=1.62.0",
    "protobuf>=5.26.0",
    "anthropic>=0.49.0",
    "openai>=1.30.0",
    "httpx>=0.27.0",
    "structlog>=24.1.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.0",
]

[dependency-groups]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "types-protobuf>=5.26.0",
]

[project.scripts]
sourcebridge-worker = "workers.__main__:main"

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM"]

[tool.mypy]
python_version = "3.12"
warn_unused_configs = true
check_untyped_defs = true
# Relax strict checks that conflict with gRPC stubs and third-party SDK types
disable_error_code = ["union-attr", "arg-type", "return-value", "attr-defined"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = [".", "../gen/python"]
testpaths = [".", "../tests"]
norecursedirs = ["fixtures", "node_modules", ".git", "__pycache__"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["workers"]
